src/requirements.txt
future-map==0.1.2
gradio==4.20.1
imagehash==4.3.1
lru-dict==1.3.0
natsort==8.4.0
numpy==1.26.4
oshash==0.1.1
pandas==2.2.1
pywebview==4.4.1
safetensors==0.4.2
timm==0.9.16
torch==2.0.1
tqdm==4.66.2
